Fix 32 review findings across SAR module (3 critical, 10 high, 13 medium, 6 low)
- Add Numba JIT acceleration for phase unwrapping (array-based binary heap) and
  Goldstein filter (batch FFT + compiled overlap-add accumulation)
- Fix InSAR sign convention documentation (positive = toward sensor)
- Add memory guards before large allocations (phase unwrap, Goldstein, SBAS, coregistration)
- Replace np.mgrid with broadcasting in coregistration (saves ~4.8 GB)
- Use in-place operations in SBAS (saves ~24 GB on 20-epoch stacks)
- Fix GUI race condition with _stop_requested flag
- Add multi-sensor adapter detection in bathymetry pipeline
- Narrow broad except clauses in sensor adapters
- Add input validation, NaN handling, config versioning
- Fix circular import in bathymetry_pipeline.py (deferred imports)

+++ ocean_rs/sar/insar/interferogram.py
@@ -78,8 +78,8 @@ def form_interferogram(
     window_size = (coh_window_azimuth, coh_window_range)
 
     # Numerator: |<primary · conj(secondary)>|
-    cross_mean = uniform_filter(ifg.real, size=window_size) + \
-                 1j * uniform_filter(ifg.imag, size=window_size)
+    cross_mean = uniform_filter(ifg.real.astype(np.float64), size=window_size) + \
+                 1j * uniform_filter(ifg.imag.astype(np.float64), size=window_size)
     numerator = np.abs(cross_mean)
 
     # Denominator: √(<|primary|²> · <|secondary|²>)
